requirements.txt
Django==3.2.4

# Server/Provider
django-oauth-toolkit==1.5.0  # https://github.com/jazzband/django-oauth-toolkit
django-cors-headers==3.7.0  # https://github.com/adamchainz/django-cors-headers

# Web Consumer
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
requests==2.25.1  # https://github.com/psf/requests

# Mobile Consumer (requires the consumer requirements above as well)
djangorestframework==3.12.4
# simplejwt is a handy tool for authentication using mobile apps
# Take a look at https://github.com/Andrew-Chen-Wang/mobile-auth-example
# for an implementation example for Swift and Kotlin (native lang for mobile OS's)
djangorestframework-simplejwt==4.7.1
